Fix test directory __init__ and create layer tests
Corrected the __init__ for testing. Created tests for the BaseLayer,
Source, and Terminator classes in the layer module. These tests check
for the default value for some attributes, then change those attributes
and make sure we get values we expect.

Mostly this was an experiment in using the unittest package.