Fix variable references and imports; add TODO
Fixed references to 'polarization' which should now be 'pol' everywhere.
Also added a numpy import within the module in the spirit of 'explicit
is better than implicit'.

Created a new TODO in the module-level docstring. Realized that we can
get weird behavior from trig functions at angles in the vicinity of
pi/2. That will need to be handled more carefully than it currently is.
This use case is not a super high priority because we should know a
priori that non-incident rays should not reflect or transmit, so this
doesn't really hurt functionality.

But I should fix it in the interest of not being intentionally sloppy.