Create core.main(); minor fixes throughout core.py
Created the main function of the core module. The function takes a
dictionary of parameters as input, performs the core
transmission/reflection calculations in the correct order, and returns a
dictionary of results.

Minor fixes include correcting references within rt_amp() and
eliminating the 'lossy' argument in wavenumber(). It is easy to pass an
array of zeros as the loss tangents if we want to run a lossless
simulation. This option was removed in the spirit of explicit >>
implicit.