Add option to calculate T with freq-dep loss
Created a couple new functions and checks to handle a
frequency-dependent loss parameter as defined by the Halpern 'a' and 'b'
coefficients. If a layer object is created with 'a' and 'b'
coefficients, then the main function will calculate an equivalent loss
tangent for it at each frequency. This will override any loss tangent
that was previously set for that layer.

This is carried out in two functions: one that caclulates the imaginary
part of the refractive index (this may be used in the future) and one
that calculates the loss tangent from the imaginary part of the
refractive index.

Results agree with the Halpern method to < 2% error over 500 GHz, which
is close enough for our purposes.