Create check for Halpern coeff's in model set-up
Added a check to catch layers that have been instantiated with Halpern
coefficients. These layers are parsed for position in the assembly
(i.e., where it falls in the layer stack---i.e, i.e, it's list index),
'halperna' and 'halpernb' coefficients, and the layer's refractive index
(the real part).

This information is stored in a dictionary keyed by the layer's list
index, and then passed to the main simulation function.