Fix layer __repr__ in output header
Model.struct contained the wrong string representations of the objects
it contained. Source and Terminator layers were properly recorded, but
and user-defined layer was given 'Basic layer' status whether a custom
description was set or not. This is now corrected. User defined layers
that are part of the simulation are flagged as 'Sim layer' and custom
descriptions are stored properly if they are passed.

The header is pretty unwieldly so it may be worth writing the header to
the file in a somewhat more readable way. After all, if someone is
looking at the file, it's probably because they're trying to read the
header.