Check for arbitrary terminator refractive index
Added a check to allow an arbitrary refractive index for the terminating
layer. If the user passes a refractive index that is not equal to one,
then the problemis set up as normal. If the user passes a refractive
index that is equal to one, then we check for whether Terminator.vac is
True and handle the setting the refractive index like we did before.