Allow Source/Terminator arbitrary n/tand; update docstrings
This commit is not as atomic as I had intended due to user error on my
part. Oh well. This amended commit message will have to do.

The main functionality added in this commit is to allow the user to pass
an arbitrary value for the refractive index and loss tangent when
instantiating the Source and Terminator objects. To this end, I have
included new information in the Source and Terminator class-level
docstrings. The Source layer docstring now warns the user against setting
a non-zero loss tangent.