Refactor test_core.py to implement pytest
I've decided to use pytest instead of unittest, in part because this
will force me to write new tests and in part because I want to test
drive pytest a little more.

Basically, I've burned down the old tests and started over with fresh
eyes. All but a couple functions in `core.py` are now being tested by
`test_core.py`. The last two functions, which are flagged to be skipped
by pytest, will be easier to check in an integration test.

All tests flagged to run are passing.