Write tests for `armm/model.py`
I've included tests for all functions in `model.py`. All but one
function is tested directly. The odd one out (`_set_up_sim`) is tested
through `model.run` and `model.set_up`.

When testing `model.save`, I create an output file. It's impolite to
leave this thing hanging around, so I'm deleting it and the end of the
tests.