Add a requirements.txt
This should make it easier to maintain the dependencies, especially if I
add visualization tools. Right now dependencies are pegged at the
version number on my own machine (numpy>=1.16 and scipy>=1.3), but I am
pretty surethe package runs with earlier versions. Will update the
version requirements when this is confirmed.